+++ CHANGELOG.md
@@ -18,6 +18,7 @@ key binding with `play_from_formats` command.
 ### Fixed
 - Don't block when resolving channel ids.
 - Make buildable on non unix systems.
+- Add new videos in the correct order.
 
 ## [0.6.0] - 2025-06-29
 ### Added